Struct nsstring_vendor::nsCStringRepr

#[repr(C)]

pub struct nsCStringRepr { /* fields omitted */ }

The representation of a ns[C]String type in C++. This type is used internally by our definition of ns[C]String to ensure layout compatibility with the C++ ns[C]String type.

This type may also be used in place of a C++ ns[C]String inside of struct definitions which are shared with C++, as it has identical layout to our ns[C]String type. Due to drop flags, our ns[C]String type does not have identical layout. When drop flags are removed, this type will likely be made a private implementation detail, and its uses will be replaced with ns[C]String.

This struct will leak its data if dropped from rust. See the module documentation for more information on this type.